+++ openmetadata-ui/src/main/resources/ui/src/components/AuditLog/AuditLogList.less
@@ -48,10 +48,6 @@
     border-bottom: none;
   }
 
-  &:hover {
-    background-color: @grey-50;
-  }
-
   &.skeleton-item {
     .skeleton-avatar {
       width: 32px;
